cleanups and folder rename
